+++ demisto_sdk/common/hook_validations/structure.py
@@ -61,6 +61,7 @@ class StructureValidator:
         if self.old_file:  # In case the file is modified
             answers.append(not self.is_id_modified())
             answers.append(self.is_valid_fromversion_on_modified())
+
         return all(answers)
 
     def scheme_of_file_by_path(self):
@@ -70,6 +71,7 @@ class StructureValidator:
         Returns:
             (str): Type of file by scheme name
         """
+
         for scheme_name, regex_list in SCHEMA_TO_REGEX.items():
             if get_matching_regex(self.file_path, regex_list):
                 return scheme_name
@@ -90,7 +92,7 @@ class StructureValidator:
         Returns:
             bool. Whether the scheme is valid on self.file_path.
         """
-        if self.scheme_name in [None, 'reputation']:
+        if self.scheme_name in [None, 'reputation', 'image']:
             return True
         try:
             path = os.path.normpath(
@@ -156,7 +158,12 @@ class StructureValidator:
 
         old_version_id = self.get_file_id_from_loaded_file_data(self.old_file)
         new_file_id = self.get_file_id_from_loaded_file_data(self.current_file)
-        return not (new_file_id == old_version_id)
+        if not (new_file_id == old_version_id):
+            print_error(f"The file id for {self.file_path} has changed from {old_version_id} to {new_file_id}")
+            return True
+
+        # False - the id has not changed.
+        return False
 
     def is_valid_fromversion_on_modified(self):
         # type: () -> bool
@@ -191,6 +198,10 @@ class StructureValidator:
                 with open(self.file_path, 'r') as file_obj:
                     loaded_file_data = load_function(file_obj)  # type: ignore
                     return loaded_file_data
+
+            # Ignore loading image
+            elif file_extension == '.png':
+                return {}
 
         print_error(Errors.wrong_file_extension(file_extension, self.FILE_SUFFIX_TO_LOAD_FUNCTION.keys()))
         return {}