+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 0.3.4
 * Fixed an issue where "_test" file for scripts/integrations created using **init** would import the "HelloWorld" templates.
 * Add validation for docker image inside integrations and scripts.
+* Add --use-git flag to format command to format all changed files.
 
 ### 0.3.3
 * Added backwards compatibility break error message.